@@ -35,36 +35,81 @@
     "publish:patch": "npm version patch && npm run prepack && npm publish build",
     "setup": "npm install && typings install",
     "typings": "typings",
-    "start": "php -S localhost:8000 -t dist/ local_dev/router.php"
+    "start": "node server.js"
   },
   "dependencies": {
-    "browserify-fs": "^1.0.0",
+    "@types/history": "^2.0.48",
+    "@types/jasmine": "^2.5.47",
+    "@types/jwt-decode": "^1.4.28",
+    "@types/material-ui": "^0.17.4",
+    "@types/react": "^15.0.23",
+    "@types/react-dom": "^15.5.0",
+    "@types/react-hot-loader": "^3.0.1",
+    "@types/react-redux": "^4.4.40",
+    "@types/react-router": "^2.0.49",
+    "@types/react-tap-event-plugin": "0.0.30",
+    "@types/redux-form": "^6.6.2",
+    "@types/redux-thunk": "^2.1.0",
+    "@types/webpack-env": "^1.13.0",
+    "autoprefixer": "^6.3.6",
+    "babel-core": "^6.17.0",
+    "babel-loader": "^6.2.5",
+    "babel-preset-es2015": "^6.14.0",
+    "babel-preset-react": "^6.11.1",
+    "babel-preset-stage-0": "^6.3.13",
+    "babel-runtime": "^6.11.6",
+    "compression": "^1.6.2",
+    "css-loader": "^0.28.0",
+    "cssnano": "^3.7.4",
+    "debug": "^2.2.0",
     "express": "^4.15.2",
+    "extract-text-webpack-plugin": "^2.1.0",
+    "file-loader": "^0.9.0",
+    "fs-extra": "^1.0.0",
     "geopattern": "^1.2.3",
+    "html-webpack-plugin": "^2.22.0",
+    "imports-loader": "^0.7.0",
+    "json-loader": "^0.5.4",
     "jwt-decode": "^2.1.0",
-    "material-ui": "^0.16.4",
+    "material-ui": "^0.18.0",
     "moment": "^2.15.1",
     "moment-timezone": "^0.5.7",
+    "node-sass": "^4.0.0",
+    "normalize.css": "^5.0.0",
     "popsicle": "^8.2.0",
-    "react": "^15.4.2",
+    "postcss-loader": "^1.1.0",
+    "prop-types": "^15.5.6",
+    "react": "^15.5.0",
     "react-addons-css-transition-group": "^15.3.2",
-    "react-data-grid": "^2.0.32",
-    "react-dom": "^15.3.2",
+    "react-data-grid": "^2.0.33",
+    "react-dom": "^15.5.4",
     "react-dropzone": "^3.7.3",
     "react-redux": "^4.4.5",
     "react-router": "^2.8.1",
     "react-tap-event-plugin": "^2.0.0",
     "redux": "^3.6.0",
     "redux-form": "^6.5.0",
-    "redux-thunk": "^2.1.0"
+    "redux-thunk": "^2.1.0",
+    "sass-loader": "^4.0.0",
+    "style-loader": "^0.13.1",
+    "ts-loader": "^2.0.3",
+    "typescript": "^2.1.6",
+    "url-loader": "^0.5.6",
+    "webpack": "^2.4.1"
   },
   "devDependencies": {
-    "browserify": "^13.1.0",
+    "awesome-typescript-loader": "^3.1.2",
+    "babel-plugin-istanbul": "^3.0.0",
     "istanbul": "^0.4.5",
     "jasmine": "^2.5.2",
     "npm-run-all": "^3.1.0",
-    "typescript": "^2.0.3",
-    "typings": "^1.4.0"
+    "react-hot-loader": "^3.0.0-beta.6",
+    "resolve-url-loader": "^2.0.2",
+    "source-map-loader": "^0.2.1",
+    "typescript": "^2.2.1",
+    "webpack-dev-middleware": "^1.6.1",
+    "webpack-dev-server": "^2.4.5",
+    "webpack-hot-middleware": "^2.12.2"
   },
   "main": "index.js"
 }
